requirements.txt
requests
python-dotenv
opencv-python
winsdk
colorama
keyring
pyinstaller
